requirements.txt
aiogram
numpy
pdfkit
Pillow
apscheduler
